requirements.txt
aeneas==1.7.3.0
aioredis==1.2.0
alabaster==0.7.8
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.1.0
audioread==2.1.9
autobahn==18.10.1
Automat==0.7.0
autopep8==1.5.4
Babel==2.4.0
beautifulsoup4==4.10.0
boto3==1.14.31
botocore==1.17.31
Brlapi==0.6.6
cachelib==0.1
cachetools==4.0.0
certifi==2020.4.5.1
cffi==1.15.0
channels==2.2.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==36.0.1
cs50==5.0.4
cupshelpers==1.0
daphne==2.3.0
dataclasses==0.8
decorator==5.1.1
defer==1.0.6
defusedxml==0.5.0
Deprecated==1.2.13
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.4
django-allauth==0.38.0
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-extensions==3.1.5
django-filter==2.3.0
django-js-asset==1.2.2
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-sslserver==0.22
django-storages==1.9.1
django-utils-six==2.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.6
dnspython==2.2.1
docopt==0.6.2
docutils==0.15.2
ecdsa==0.17.0
enum34==1.1.10
Faker==13.2.0
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
gitsome==0.7.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.0.4
h5py==2.10.0
hiredis==0.2.0
httplib2==0.20.4
hyperlink==18.0.0
idna==2.9
imagesize==0.7.1
importlib-metadata==4.8.1
importlib-resources==5.3.0
incremental==17.5.0
install==1.3.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.1.0
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lark-parser==0.12.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
librosa==0.9.1
llvmlite==0.36.0
louis==3.5.0
lxml==4.7.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
meld==3.18.0
mongoengine==0.24.0
msgpack==0.6.2
mysqlclient==2.1.0
netifaces==0.10.4
numba==0.53.1
numpy==1.19.5
numpydoc==0.7.0
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
opentok==3.2.0
packaging==21.3
pandas==1.0.3
passlib==1.7.4
pbr==5.8.0
pexpect==4.2.1
Pillow==8.0.1
ply==3.11
pooch==1.6.0
prompt-toolkit==1.0.15
protobuf==3.19.4
psycopg2==2.8.5
psycopg2-binary==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydub==0.25.1
Pygments==2.2.0
PyGObject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.12.3
PyMySQL==1.0.2
PyNaCl==1.1.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.13.0
python-jose==3.3.0
python3-openid==3.1.0
pytube3==9.6.4
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
redis==4.0.2
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
required==0.4.0
resampy==0.2.2
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==21.1.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.0
Sphinx==1.6.7
SQLAlchemy==1.3.17
sqlparse==0.2.4
ssh-import-id==5.7
stripe==2.70.0
system-service==0.3
systemd-python==234
termcolor==1.1.0
testresources==2.0.1
textdistance==4.2.2
toml==0.10.2
tqdm==4.62.3
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.9
usb-creator==0.3.3
virtualenv==20.2.1
wadllib==1.3.2
wcwidth==0.1.7
websocket-client==1.2.3
Werkzeug==0.15.4
whitenoise==5.1.0
wrapt==1.12.1
xkit==0.0.0
XlsxWriter==3.0.2
xonsh==0.6.0
yapf==0.31.0
youtube-dl==2021.12.17
youtube-transcript-api==0.4.3
zipp==3.6.0
zope.interface==4.6.0
